Update backend controllers, services and frontend UI

+++ src/main/java/com/railway/concessionsystem/repository/StaffRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StaffRepository extends JpaRepository<Staff, Long> {
+public interface StaffRepository extends JpaRepository<Staff, Integer> { // Changed from Long to Integer ✅
     Optional<Staff> findByEmail(String email);
     boolean existsByEmail(String email);
 }
